requirements.txt
# requirements.txt – All dependencies for AI Travel Planner

# Core packages
fastapi>=0.115.0
uvicorn>=0.34.0
streamlit==1.33.0

# Google Gemini & ADK
google-generativeai==0.5.2
google-adk==0.4.0

# Tools
httpx>=0.28.1,<1.0.0
pydantic>=2.7.2,<3.0.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0